feat: creates main-testimonials component
